Replace Route.hander with Route.handler_class
This means that handler cannot be determined on a Route by Route basis,
but I can't remember why I thought that was a good idea anyway!

Replacing the dotted path with a concrete class opens the posibility of
adding model checks for particular handlers. That'll come in another
commit soon.