Allow django admin configuration
When using django's admin, it's now important to add the
`CONMAN_ADMIN_ROUTES` setting. This is a list (or tuple) of dotted paths
to models as accepted by `django.apps.apps.get_model`. Eg, for a
hypothetical `MyModel` subclass of `Route` in a `my_app` app, we would
use the string `my_app.MyModel`.

There are still a few rough edges, and the documentation is woefully
lacking, but it'd be really nice to have this feature.

Verbose names and help texts on fields on models should also help to
make the admin a slightly nicer place to be.

Fixes #114 and #79.